requirements.txt
txtai
simple_term_menu
pyipv8